+++ server/listeners/messages/direct-message.ts
@@ -1,5 +1,6 @@
 import type { AllMiddlewareArgs, SlackEventMiddlewareArgs } from "@slack/bolt";
 import { type ChatDetail, v0 } from "v0-sdk";
+import { extractV0Summary } from "~/lib/ai/utils";
 import {
   getLastAssistantMessage,
   getThreadContextAsModelMessage,
@@ -7,48 +8,12 @@ import {
   updateAgentStatus,
 } from "~/lib/slack/utils";
 
-/**
- * Extracts the final summary text from v0's structured response format
- * Removes <Thinking>, <V0LaunchTasks>, <CodeProject> tags and returns clean summary
- */
-function extractV0Summary(rawContent: string): string {
-  if (!rawContent) return "";
-
-  // Remove <Thinking> blocks
-  let cleaned = rawContent.replace(/<Thinking>[\s\S]*?<\/Thinking>/gi, '');
-
-  // Remove <V0LaunchTasks> blocks
-  cleaned = cleaned.replace(/<V0LaunchTasks>[\s\S]*?<\/V0LaunchTasks>/gi, '');
-
-  // Remove <CodeProject> blocks (including self-closing and multi-line variants)
-  cleaned = cleaned.replace(/<CodeProject[^>]*>[\s\S]*?<\/CodeProject>/gi, '');
-  cleaned = cleaned.replace(/<CodeProject[^>]*\/>/gi, '');
-
-  // Remove any remaining XML-like tags that might be left
-  cleaned = cleaned.replace(/<\/?[A-Za-z][A-Za-z0-9]*[^>]*>/g, '');
-
-  // Clean up extra whitespace and newlines
-  cleaned = cleaned.trim();
-  cleaned = cleaned.replace(/\n\s*\n/g, '\n'); // Remove empty lines
-  cleaned = cleaned.replace(/^\s+|\s+$/gm, ''); // Trim each line
-
-  // If there are multiple paragraphs, take the last meaningful one
-  const paragraphs = cleaned.split('\n').filter(p => p.trim().length > 0);
-
-  if (paragraphs.length === 0) {
-    return "Task completed successfully.";
-  }
-
-  // Return the last substantial paragraph (usually the summary)
-  return paragraphs[paragraphs.length - 1];
-}
-
 export const directMessageCallback = async ({
   message,
   say,
   logger,
   context,
-  client
+  client,
 }: AllMiddlewareArgs & SlackEventMiddlewareArgs<"message">) => {
   // @ts-expect-error
   const { channel, thread_ts, text } = message;
@@ -103,7 +68,9 @@ export const directMessageCallback = async ({
         .then((chat: ChatDetail) => {
           v0ChatId = chat.id;
           demoUrl = chat.latestVersion.demoUrl;
-          summary = extractV0Summary(chat.messages[chat.messages.length - 1].content);
+          summary = extractV0Summary(
+            chat.messages[chat.messages.length - 1].content,
+          );
 
           client.assistant.threads.setTitle({
             channel_id: channel,
@@ -116,13 +83,17 @@ export const directMessageCallback = async ({
       if (lastAssistantMessage?.metadata?.event_payload?.chat_id) {
         v0ChatId = lastAssistantMessage.metadata.event_payload.chat_id;
       }
-      await v0.chats.sendMessage({
-        chatId: v0ChatId,
-        message: text,
-      }).then((message: ChatDetail) => {
-        summary = extractV0Summary(message.messages[message.messages.length - 1].content);
-        demoUrl = message.latestVersion.demoUrl;
-      });
+      await v0.chats
+        .sendMessage({
+          chatId: v0ChatId,
+          message: text,
+        })
+        .then((message: ChatDetail) => {
+          summary = extractV0Summary(
+            message.messages[message.messages.length - 1].content,
+          );
+          demoUrl = message.latestVersion.demoUrl;
+        });
     }
 
     await say({
